Move utility methods in Experience component to helpers.ts file
